Add files via upload
I'm older than I was 50 but it's not easy but I have Jesus in my life and if I stay single for the rest of my life I'll live with it. My father said one day id be alone in the world and he was right. He left the country back on October 3 1983